server/requirements.txt
alembic==1.7.4
astroid==2.8.4
async-timeout==4.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cognitojwt==1.4.1
cryptography==36.0.1
Deprecated==1.2.13
ecdsa==0.17.0
Flask==2.0.2
Flask-Cognito==1.18
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
greenlet==1.1.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jwt==1.3.1
lazy-object-proxy==1.6.0
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.21
pylint==2.11.1
pyparsing==3.0.8
python-dotenv==0.19.2
python-jose==3.3.0
redis==4.2.2
requests==2.27.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
